+++ NEWS.md
@@ -1,7 +1,10 @@
 # soilDB 2.7.9 (2023-09-01)
  
- - Added new `method="extent"` option for `fetchSDA_spatial()`. This method calculates a bounding box for each `mukey` or `nationalmusym` extent. 
-  - This contrasts with existing `method="bbox"` which returns a rectangular bounding box for each delineation. Mapunits used on multiple legends, having multiple `mukey` are concatenated into comma separated list for the result.
+ - Added new `method` options for `fetchSDA_spatial()`. Aggregation grouping is controlled by the `by.col` argument. This works for mapunit and survey area polygon geometries, aggregating all polygons in the group for each `mukey`, `nationalmusym`, `lkey`, or `areasymbol` extent.
+  - `method="extent"` method calculates a bounding rectangle  
+  - `method="convexhull"` calculates the convex hull 
+  - `method="union"` returns a MULTIPOLYGON 
+  - `method="collection"` returns a GEOMETRYCOLLECTION 
  
 ## Bug Fixes